+++ src/translators/crate.py
@@ -7,7 +7,6 @@ from translators import base_translator
 from utils.common import iter_entity_attrs
 import logging
 import os
-import statistics
 from geocoding.slf import SlfQuery
 from .crate_geo_query import from_ngsi_query
 
@@ -48,7 +47,7 @@ METADATA_TABLE_NAME = "md_ets_metadata"
 FIWARE_SERVICEPATH = 'fiware_servicepath'
 TENANT_PREFIX = 'mt'
 TYPE_PREFIX = 'et'
-VALID_AGGR_METHODS = ['count', 'sum', 'avg', 'min', 'max',]
+VALID_AGGR_METHODS = ['count', 'sum', 'avg', 'min', 'max']
 VALID_AGGR_PERIODS = ['year', 'month', 'day', 'hour', 'minute', 'second']
 
 
@@ -314,10 +313,9 @@ class CrateTranslator(base_translator.BaseTranslator):
                         values.append([float(lon), float(lat)])
                     else:
                         values.append(e[cn]['value'])
-                except KeyError as e:
-                    msg = ("Seems like not all entities of same type came "
-                           "with the same set of attributes. {}").format(e)
-                    raise NotImplementedError(msg)
+                except KeyError:
+                    # this entity update does not have a value for the column so use None which will be inserted as NULL to the db.
+                    values.append( None )
         return values
 
 
@@ -351,7 +349,7 @@ class CrateTranslator(base_translator.BaseTranslator):
         else:
             # Bring translation table!
             stmt = "select entity_attrs from {} where table_name = ?"
-            self.cursor.execute(stmt.format(METADATA_TABLE_NAME), [table_name,])
+            self.cursor.execute(stmt.format(METADATA_TABLE_NAME), [table_name])
 
             # By design, one entry per table_name
             res = self.cursor.fetchall()
@@ -374,6 +372,7 @@ class CrateTranslator(base_translator.BaseTranslator):
         if fiware_service:
             where = "where table_name ~* '\"{}{}\"[.].*'"
             op += where.format(TENANT_PREFIX, fiware_service.lower())
+
         try:
             self.cursor.execute(op)
         except exceptions.ProgrammingError as e:
@@ -634,7 +633,9 @@ class CrateTranslator(base_translator.BaseTranslator):
         if entity_id:
             entity_ids = tuple([entity_id])
 
-        select_clause = self._get_select_clause(attr_names,
+        lower_attr_names = [a.lower() for a in attr_names]\
+            if attr_names else attr_names
+        select_clause = self._get_select_clause(lower_attr_names,
                                                 aggr_method,
                                                 aggr_period)
         if not where_clause:
@@ -740,10 +741,17 @@ class CrateTranslator(base_translator.BaseTranslator):
         self.cursor.execute(stmt, [table_name])
         res = self.cursor.fetchall()
 
+        if len(res) == 0:
+            # See GH #173
+            tn = ".".join([x.strip('"') for x in table_name.split('.')])
+            self.cursor.execute(stmt, [tn])
+            res = self.cursor.fetchall()
+
         if len(res) != 1:
             msg = "Cannot have {} entries in table '{}' for PK '{}'"
             msg = msg.format(len(res), METADATA_TABLE_NAME, table_name)
             self.logger.error(msg)
+            raise RuntimeError(msg)
 
         entities = {}
         entity_attrs = res[0][0]
@@ -760,8 +768,9 @@ class CrateTranslator(base_translator.BaseTranslator):
 
                 # CrateDBs and NGSI use different geo:point coordinates order.
                 if original_type == NGSI_GEOPOINT:
-                    lon, lat = v
-                    v = "{}, {}".format(lat, lon)
+                    if v is not None:
+                        lon, lat = v
+                        v = "{}, {}".format(lat, lon)
 
                 if original_name in (NGSI_TYPE, NGSI_ID):
                     e[original_name] = v
@@ -850,10 +859,17 @@ class CrateTranslator(base_translator.BaseTranslator):
         # Delete entry from metadata table
         op = "delete from {} where table_name = ?".format(METADATA_TABLE_NAME)
         try:
-            self.cursor.execute(op, [table_name,])
+            self.cursor.execute(op, [table_name])
         except exceptions.ProgrammingError as e:
-            # What if this one fails and previous didn't?
             logging.debug("{}".format(e))
+
+        if self.cursor.rowcount == 0 and table_name.startswith('"'):
+            # See GH #173
+            old_tn = ".".join([x.strip('"') for x in table_name.split('.')])
+            try:
+                self.cursor.execute(op, [old_tn])
+            except exceptions.ProgrammingError as e:
+                logging.debug("{}".format(e))
 
         return count
 
@@ -871,9 +887,11 @@ class CrateTranslator(base_translator.BaseTranslator):
         if fiware_service is None:
             wc = "where table_name NOT like '\"{}%.%'".format(TENANT_PREFIX)
         else:
-            # See _et2tn
-            prefix = '"{}{}"'.format(TENANT_PREFIX, fiware_service.lower())
-            wc = "where table_name like '{}.%'".format(prefix)
+            # Old is prior QL 0.6.0. GH #173
+            old_prefix = '{}{}'.format(TENANT_PREFIX, fiware_service.lower())
+            prefix = self._et2tn("FooType", fiware_service).split('.')[0]
+            wc = "where table_name like '{}.%' " \
+                 "or table_name like '{}.%'".format(old_prefix, prefix)
 
         stmt = "select distinct(table_name) from {} {}".format(
             METADATA_TABLE_NAME,